Split out adventure related code into its own files
