+++ Assets/Scripts/Simulation/Adventures/AdventureEncounterAct.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ac00eaa0b981554c865efa94418a521
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
