+++ Assets/Scripts/Simulation/Adventures/AdventureEncounterPath.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 510b7ca7f5a0338409a99769750ea5db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
