+++ Assets/Scripts/Simulation/Adventures/IAdventureContextCriteria.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aced658c5d6ca2645a82c7878475bf98
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
